Move PdfHandler code to view model
