+++ app/src/main/java/com/michaeltroger/gruenerpass/update/AppMigrateFrom6.kt
@@ -1,7 +1,7 @@
 package com.michaeltroger.gruenerpass.update
 
 import android.content.Context
-import com.michaeltroger.gruenerpass.pdf.PDF_FILENAME
+import com.michaeltroger.gruenerpass.PDF_FILENAME
 import java.io.File
 import java.lang.Exception
